Avoid Noodle Dimming by selecting nodes
Noodle gets dimmed if none of the nodes on its both end are visible or selected.
Noodle does not get dimmed if one of the node on its end is selected.
So to avoid noodle dimming we just select all the nodes.
But active node and selected nodes have distinctive outline. We change the outline colors in Prefereces and restore it afte the screenshot.
Currently setting the outline color to black which works well with default blender theme. If user has custom theme, there's a option provided in addonn preferces to customize the outline color while taking screenshots.